Tracking topic28

+++ tracker/tracking/components/Rectangle.py
@@ -1,7 +1,7 @@
 from components.Point2i import Point2i
 
 class Rectangle():
-    def __init__(self, upperLeftPoint:Point2i, width:int, height:int):
+    def __init__(self, upperLeftPoint:Point2i, width:float, height:float):
         self.UpperLeftPoint:Point2i = upperLeftPoint
-        self.Width:int = width
-        self.Height:int = height
+        self.Width:float = width
+        self.Height:float = height
